Attached the label rendering logic to a later (post-plugin-initialization) event, and removed our Twig extension in favor of a simple global pass-through to the Label service.
This makes a bit more functionality accessible to templates, and it prevents our plugin from trying to use Twig [to render the label text] before other plugins have had a chance to initialize. (If another plugin tried to register a Twig extension after we had already used Twig to render our template, there'd be collisions and error messages and weeping and gnashing of teeth.)

Updated version to 3.1.1.